+++ e2e/mocks/types.ts
@@ -1,0 +1,3 @@
+export type BBNWalletType = "Keplr" | "Leap" | "Cosmostation";
+
+export type BTCWalletType = "OKX" | "Unisat" | "OneKey" | "Keystone";
